requirements.txt
python-bittrex
python-binance
requests
